feat(switch): update acquiring fee client, merchant acquirer integration, update route service and logger tests

+++ services/switch/src/main/java/com/processing/service/MerchantAcquirerClient.java
@@ -8,6 +8,8 @@ import org.springframework.http.HttpMethod;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClient;
 
+import java.math.BigDecimal;
+
 @Service
 public class MerchantAcquirerClient implements AcquiringFeeClient {
 
@@ -22,7 +24,7 @@ public class MerchantAcquirerClient implements AcquiringFeeClient {
     }
 
     @Override
-    public Long fetchAcquiringFee(String transmissionDateTime, String stan, String pan, String terminalId, Long amount) {
+    public BigDecimal fetchAcquiringFee(String transmissionDateTime, String stan, String pan, String terminalId, BigDecimal amount) {
         try {
             AcquirerFeeResponse response = restClient.method(HttpMethod.GET)
                     .uri(switchProperties.merchantAcquirerUrl() + "/api/simulator/merchant/fee")
@@ -44,12 +46,12 @@ public class MerchantAcquirerClient implements AcquiringFeeClient {
             String stan,
             String pan,
             @JsonProperty("terminalId") String terminalId,
-            Long amount
+            BigDecimal amount
     ) {
     }
 
     private record AcquirerFeeResponse(
-            Long acquirerFee
+            BigDecimal acquirerFee
     ) {
     }
 }
